requirements.txt
alembic==1.7.4
click==8.0.3
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.1
importlib-resources==5.3.0
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2==2.9.1
python-dotenv==0.19.1
SQLAlchemy==1.4.26
Werkzeug==2.0.2
zipp==3.6.0
